obs-rc/requirements.txt
obs-websocket-py
fastapi
uvicorn
requests